implement prompt:none flow
Implement prompt:none flow to match that proposed in
uisautomation/hydra-consent-app#5. This commit is intended to keep the
mock consent app and real consent app in lock step with regards to
features.

Update the app to cache the current subject in the flask session and
automatically grant requests if there is a user in the session. This
approximates the existing Raven behaviour.

Teach the app about the "prompt:none" scope which signals that the
request should be immediately rejected if there is not a user in the
session.